+++ opentelemetry-etw-metrics/src/exporter/mod.rs
@@ -70,14 +70,25 @@ impl PushMetricsExporter for MetricsExporter {
             .encode(&mut byte_array)
             .map_err(|err| MetricsError::Other(err.to_string()))?;
 
-        let result = etw::write(&byte_array);
-        if result != 0 {
+        if (byte_array.len()) > etw::MAX_EVENT_SIZE {
             global::handle_error(MetricsError::Other(format!(
-                "Failed to write ETW event with error code: {}",
-                result
+                "Exporting failed due to event size {} exceeding the maximum size of {} bytes",
+                byte_array.len(),
+                etw::MAX_EVENT_SIZE
             )));
+        } else {
+            let result = etw::write(&byte_array);
+            // TODO: Better logging/internal metrics needed here for non-failure
+            // case Uncomment the line below to see the exported bytes until a
+            // better logging solution is implemented
+            // println!("Exported {} bytes to ETW", byte_array.len());
+            if result != 0 {
+                global::handle_error(MetricsError::Other(format!(
+                    "Failed to write ETW event with error code: {}",
+                    result
+                )));
+            }
         }
-
         Ok(())
     }